+++ model/team_org.go
@@ -5,9 +5,9 @@ type TeamOrgs []*TeamOrg
 
 // TeamOrg represents a team org model definition.
 type TeamOrg struct {
-	TeamID int    `json:"team_id" sql:"index"`
+	TeamID int64  `json:"team_id" sql:"index"`
 	Team   *Team  `json:"team,omitempty"`
-	OrgID  int    `json:"org_id" sql:"index"`
+	OrgID  int64  `json:"org_id" sql:"index"`
 	Org    *Org   `json:"org,omitempty"`
 	Perm   string `json:"perm,omitempty"`
 }
